fix: 修复 USB 事件信号发射中的类型转换问题

+++ examples/Qt-usb-monitor.cpp
@@ -9,7 +9,7 @@ public:
     UsbWatcher() {
         monitor_ = std::make_unique<USBMonitor>(
             [this](UsbState s, std::string p) {
-                emit usbEvent(s, QString::fromStdString(p));
+                emit usbEvent(static_cast<int>(s), QString::fromStdString(p));
             }
         );
         monitor_->startMonitoring();
